+++ packages/com.eframework.ui-extras/Runtime/UIHelper/ToggleHelper.cs
@@ -1,6 +1,3 @@
-﻿using DG.Tweening;
-using EFramework.Runtime.Audio;
-using System;
 using UnityEngine;
 using UnityEngine.EventSystems;
 using UnityEngine.UI;
@@ -12,15 +9,18 @@ namespace EFramework.Extensions.UI.Extras.UIHelper
     {
         public float DisableTime = 0.3f;
         private Toggle m_Toggle;
+
         private void Awake()
         {
             m_Toggle = GetComponent<Toggle>();
         }
+
         public void OnPointerClick(PointerEventData eventData)
         {
             m_Toggle.targetGraphic.raycastTarget = false;
             if (DisableTime > 0) Invoke("CallLater", DisableTime);
         }
+
         public Toggle GetToggle()
         {
             return m_Toggle;