+++ packages/com.eframework.ui-extras/Runtime/UIHelper/ToggleIconHelper.cs
@@ -1,4 +1,4 @@
-﻿using DG.Tweening;
+using EFramework.Runtime.Tween;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
@@ -18,10 +18,34 @@ namespace EFramework.Extensions.UI.Extras.UIHelper
         [SerializeField] private TextMeshProUGUI m_OnText;
         [SerializeField] private TextMeshProUGUI m_OffText;
         public Toggle Toggle => m_toggle;
+
         void Awake()
         {
             m_toggle?.onValueChanged.AddListener(OnToggleValueChanged);
             UpdateState(m_toggle.isOn, false);
+        }
+
+        private void OnDestroy()
+        {
+            if (m_grabIconImage != null)
+            {
+                EFrameTween.Kill(m_grabIconImage.rectTransform);
+            }
+
+            if (m_checkMarkImage != null)
+            {
+                EFrameTween.Kill(m_checkMarkImage);
+            }
+
+            if (m_OnText != null)
+            {
+                EFrameTween.Kill(m_OnText);
+            }
+
+            if (m_OffText != null)
+            {
+                EFrameTween.Kill(m_OffText);
+            }
         }
 
         public void SetIsOnWithoutNotify(bool isOn)
@@ -42,13 +66,20 @@ namespace EFramework.Extensions.UI.Extras.UIHelper
                 if (m_grabIconImage != null)
                 {
                     var targetPos = isOn ? m_moveAmount : -m_moveAmount;
-                    m_grabIconImage.DOKill(true);
-                    m_grabIconImage.rectTransform.DOAnchorPosX(targetPos, 0.2f).SetEase(Ease.OutQuad);
+                    var rectTransform = m_grabIconImage.rectTransform;
+                    var startPos = rectTransform.anchoredPosition;
+                    var endPos = new Vector2(targetPos, startPos.y);
+                    EFrameTween.Kill(rectTransform, true);
+                    EFrameTween.Vector2(startPos, endPos, 0.2f, value => rectTransform.anchoredPosition = value, new EFrameTweenOptions
+                    {
+                        Target = rectTransform,
+                        Ease = EFrameEase.OutQuad
+                    });
                 }
+
                 if (m_checkMarkImage != null)
                 {
-                    m_checkMarkImage.DOKill(true);
-                    m_checkMarkImage.DOFade(isOn ? 1f : 0f, 0.2f);
+                    FadeGraphic(m_checkMarkImage, isOn ? 1f : 0f, 0.2f);
                 }
             }
             else
@@ -58,23 +89,34 @@ namespace EFramework.Extensions.UI.Extras.UIHelper
                     var targetPos = isOn ? m_moveAmount : -m_moveAmount;
                     m_grabIconImage.rectTransform.anchoredPosition = new Vector2(targetPos, 0f);
                 }
+
                 if (m_checkMarkImage != null)
                 {
                     m_checkMarkImage.color = new Color(1f, 1f, 1f, isOn ? 1f : 0f);
                 }
             }
-            
+
             if (m_OnText != null)
             {
-                m_OnText.DOKill(true);
-                m_OnText.DOFade(isOn ? 1f : 0f, 0.2f);
+                FadeGraphic(m_OnText, isOn ? 1f : 0f, 0.2f);
             }
 
-            if (m_OffText != null) 
+            if (m_OffText != null)
             {
-                m_OffText.DOKill(true);
-                m_OffText.DOFade(isOn ? 0f : 1f, 0.2f);
+                FadeGraphic(m_OffText, isOn ? 0f : 1f, 0.2f);
             }
+        }
+
+        private static void FadeGraphic(Graphic graphic, float targetAlpha, float duration)
+        {
+            EFrameTween.Kill(graphic, true);
+            var startColor = graphic.color;
+            var targetColor = new Color(startColor.r, startColor.g, startColor.b, targetAlpha);
+            EFrameTween.Color(startColor, targetColor, duration, value => graphic.color = value, new EFrameTweenOptions
+            {
+                Target = graphic,
+                Ease = EFrameEase.Linear
+            });
         }
 
         void Reset()